+++ R/crew_controller.R
@@ -67,6 +67,7 @@ crew_class_controller <- R6::R6Class(
     .log = NULL,
     .error = NULL,
     .backlog = NULL,
+    .autoscaling = NULL,
     .shove = function(
       command,
       data = list(),
@@ -148,6 +149,11 @@ crew_class_controller <- R6::R6Class(
     #' @field backlog Character vector of explicitly backlogged tasks.
     backlog = function() {
       .subset2(private, ".backlog")
+    },
+    #' @field autoscaling `TRUE` or `FALSE`, whether async `later`-based
+    #'   auto-scaling is currently running
+    autoscaling = function() {
+      .subset2(private, ".autoscaling")
     }
   ),
   public = list(
@@ -188,6 +194,7 @@ crew_class_controller <- R6::R6Class(
       )
       crew_assert(private$.log, is.null(.) || is.list(.))
       crew_assert(private$.backlog, is.null(.) || is.character(.))
+      crew_assert(private$.autoscaling, is.null(.) || isTRUE(.) || isFALSE(.))
       invisible()
     },
     #' @description Check if the controller is empty.
@@ -298,6 +305,14 @@ crew_class_controller <- R6::R6Class(
       }
       invisible()
     },
+    #' @description Check whether the controller is started.
+    #' @details Actually checks whether the client is started.
+    #' @return `TRUE` if the controller is started, `FALSE` otherwise.
+    #' @param controllers Not used. Included to ensure the signature is
+    #'   compatible with the analogous method of controller groups.
+    started = function(controllers = NULL) {
+      isTRUE(.subset2(.subset2(self, "client"), "started"))
+    },
     #' @description Launch one or more workers.
     #' @return `NULL` (invisibly).
     #' @param n Number of workers to try to launch. The actual
@@ -343,6 +358,29 @@ crew_class_controller <- R6::R6Class(
     scale = function(throttle = TRUE, controllers = NULL) {
       private$.launcher$scale(demand = self$unresolved(), throttle = throttle)
       invisible()
+    },
+    #' @description Run worker auto-scaling in a private `later` loop
+    #'   every `controller$client$seconds_interval` seconds.
+    #' @param controllers Not used. Included to ensure the signature is
+    #'   compatible with the analogous method of controller groups.
+    #' @return `NULL` (invisibly).
+    autoscale = function(controllers = NULL) {
+      # Tested in tests/interactive/test-promises.R
+      # nocov start
+      if (isTRUE(private$.autoscaling)) {
+        return(invisible())
+      }
+      poll <- function() {
+        if (isTRUE(self$client$started) && isTRUE(private$.autoscaling)) {
+          self$scale(throttle = FALSE)
+          later::later(func = poll, delay = self$client$seconds_interval)
+        }
+      }
+      self$start()
+      private$.autoscaling <- TRUE
+      poll()
+      invisible()
+      # nocov end
     },
     #' @description Push a task to the head of the task list.
     #' @return Invisibly return the `mirai` object of the pushed task.
@@ -1127,40 +1165,46 @@ crew_class_controller <- R6::R6Class(
     #' @param seconds_interval Positive numeric of length 1, delay in the
     #'   `later::later()` polling interval to asynchronously check if
     #'   the promise can be resolved.
-    #' @param scale Logical of length 1,
-    #'   whether to automatically call `scale()`
-    #'   to auto-scale workers to meet the demand of the task load.
-    #'   Scaling up on `pop()` may be important
-    #'   for transient or nearly transient workers that tend to drop off
-    #'   quickly after doing little work.
-    #'   See also the `throttle` argument.
-    #' @param throttle `TRUE` to skip auto-scaling if it already happened
-    #'   within the last `seconds_interval` seconds. `FALSE` to auto-scale
-    #'   every time `scale()` is called. Throttling avoids
-    #'   overburdening the `mirai` dispatcher and other resources.
-    #' @param loop A `later` event loop to assign the background async
-    #'   task that auto-scales workers and polls for resolved tasks.
-    #'   Can come from `later::current_loop()`, `later::global_loop()`,
-    #'   `later::create_loop()`, or similar.
+    #' @param scale Deprecated on 2024-04-10 (version 0.9.1.9003)
+    #'   and no longer used. Now, `promise()` always turns on auto-scaling
+    #'   in a private `later` loop (if not already activated).
+    #' @param throttle Deprecated on 2024-04-10 (version 0.9.1.9003)
+    #'   and no longer used. Now, `promise()` always turns on auto-scaling
+    #'   in a private `later` loop (if not already activated).
     #' @param controllers Not used. Included to ensure the signature is
     #'   compatible with the analogous method of controller groups.
     promise = function(
       mode = "one",
       seconds_interval = 0.1,
-      scale = TRUE,
-      throttle = TRUE,
-      loop = later::current_loop(),
+      scale = NULL,
+      throttle = NULL,
       controllers = NULL
     ) {
       # Tested in tests/interactive/test-promises.R.
       # nocov start
+      crew_deprecate(
+        name = "scale",
+        date = "2024-04-10",
+        version = "0.9.1.9003",
+        alternative = c(
+          "none. promise() now always makes sure autoscaling is turned on."
+        ),
+        value = scale
+      )
+      crew_deprecate(
+        name = "throttle",
+        date = "2024-04-10",
+        version = "0.9.1.9003",
+        alternative = c(
+          "none. promise() now always makes sure autoscaling is turned on."
+        ),
+        value = throttle
+      )
+      self$autoscale(controllers = controllers)
       controller_promise(
         controller = self,
         mode = mode,
         seconds_interval = seconds_interval,
-        scale = scale,
-        throttle = throttle,
-        loop = loop,
         controllers = controllers
       )
       # nocov end
@@ -1297,6 +1341,7 @@ crew_class_controller <- R6::R6Class(
       private$.tasks <- list()
       private$.pushed <- NULL
       private$.popped <- NULL
+      private$.autoscaling <- FALSE
       invisible()
     }
   )