+++ R/utils_promises.R
@@ -11,9 +11,6 @@ controller_promise <- function(
   controller,
   mode,
   seconds_interval,
-  scale,
-  throttle,
-  loop,
   controllers
 ) {
   crew_assert(
@@ -29,18 +26,11 @@ controller_promise <- function(
     . > 0,
     message = "seconds_interval must be a finite positive number."
   )
-  crew_assert(
-    scale,
-    isTRUE(.) || isFALSE(.),
-    message = "'scale' must be TRUE or FALSE"
-  )
-  crew_assert(
-    throttle,
-    isTRUE(.) || isFALSE(.),
-    message = "'throttle' must be TRUE or FALSE"
-  )
   action <- function(resolve, reject) {
     poll <- function() {
+      if (!controller$started(controllers = controllers)) {
+        return(invisible())
+      }
       mode_one <- identical(mode, "one")
       ready <- if_any(
         mode_one,
@@ -50,16 +40,8 @@ controller_promise <- function(
       if (ready) {
         result <- if_any(
           mode_one,
-          controller$pop(
-            scale = scale,
-            throttle = throttle,
-            controllers = controllers
-          ),
-          controller$collect(
-            scale = scale,
-            throttle = throttle,
-            controllers = controllers
-          )
+          controller$pop(scale = FALSE, controllers = controllers),
+          controller$collect(scale = FALSE, controllers = controllers)
         )
         if_any(
           all(is.na(result$error)),
@@ -67,10 +49,7 @@ controller_promise <- function(
           reject(result$error[!is.na(result$error)][1L])
         )
       } else {
-        if (scale) {
-          .subset2(controller, "scale")(throttle = throttle)
-        }
-        later::later(poll, delay = seconds_interval, loop = loop)
+        later::later(poll, delay = seconds_interval)
       }
     }
     poll()