+++ NEWS.md
@@ -1,7 +1,8 @@
-# crew 0.9.1.9002 (development)
+# crew 0.9.1.9003 (development)
 
 * Use `.args` rather than `...` in `mirai::mirai()` to make sure arguments continue to be passed as local variables in `mirai` >= 0.13.1.9012.
-* Expose the `later` loop in `controller$promise()`.
+* Add new controller methods `autoscale()` and `started()` to facilitate different kinds of Shiny apps.
+* Deprecate the `scale` and `throttle` methods of `controller$promise()`. `promise()` now always calls `autoscale()` to make sure one and only one auto-scaling loop is running asynchronously. Auto-scaling thus continues even after the promise resolves.
 
 # crew 0.9.1